bootloader: switch to 64-bit version of fseek and ftell
Switch to 64-bit version of fseek() and ftell(), i.e.,
* _fseeki64() and _ftelli64() on MSVC
* fseeko() and ftello() elsewhere
and hide the differences behind pyi_fseek and pyi_ftell
macros.

Explicitly turn the bootloader's internal offset and size
variables (archive's package start, zlib package offset, remaining
size in file extraction functions) to unsigned 64-bit integer
type (uint64_t).

These changes will allow us to switch to using either unsigned
32-bit integers or 64-bit integers in the archive (entry offsets
and lengths), thereby extending its maximum allowed size either
to 4 GiB or beyond it.

On Windows, the long type used by fseek() and ftell() corresponds
to signed 32-bit integer type even on 64-bit platforms, so it
automatically limits us to 2 GiB archive files. Therefore, on
Windows we need 64-bit fseek/ftell even to have archive switch
to unsigned 32-bit integers (4 GiB support).

+++ bootloader/src/pyi_archive.h
@@ -20,6 +20,7 @@
 
 #include "pyi_global.h"
 #include <stdio.h>  /* FILE */
+#include <inttypes.h>  /* uint64_t */
 
 /* Types of CArchive items. */
 #define ARCHIVE_ITEM_BINARY           'b'  /* binary */
@@ -57,7 +58,7 @@ typedef struct _cookie {
 
 typedef struct _archive_status {
     FILE * fp;
-    size_t pkgstart;
+    uint64_t pkgstart;
     TOC *  tocbuff;
     TOC *  tocend;
     COOKIE cookie;
